Folder renaming continued / fixing
