+++ COMMANDS/Admin-Only/set-prefix.js
@@ -1,6 +1,6 @@
 const discord = require('discord.js')
 const guildSchema = require('../../Database/guildSchema')
-const updateCache = require('../../Events/updateCache')
+const updateCache = require('../../botEvents/updateCache')
 const config = require('../../config.json')
 
 module.exports = {
