+++ bot.js
@@ -51,10 +51,10 @@ client.cooldowns = new discord.Collection();
 
 
 // EVENT HANDLER
-const eventFiles = fs.readdirSync('./Events').filter(file => file.endsWith('.js'));
+const eventFiles = fs.readdirSync('./botEvents').filter(file => file.endsWith('.js'));
 
 for (const file of eventFiles) {
-	const event = require(`./Events/${file}`);
+	const event = require(`./botEvents/${file}`);
     // FOR ONE-TIME EVENTS
 	if (event.once) {
 		client.once(event.name, (...args) => event.execute(...args, client));
